requirements.txt
gspread
pandas
sqlite3
python-dotenv